Added ILogger interface
Added ILogger interface to allow Dependency Injection of custom loggers,
also provides a default logger (using the existing debug functionality)
which is used by default
