+++ EppLib/IDebugger.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace EppLib
+{
+    public interface IDebugger
+    {
+        void Log(byte[] bytes);
+
+        void Log(string str);
+    }
+}
+
